test(setup): make setup-run suite actually execute — rename off the silent-skip filename, rebuild prompt queues for the real wizard flow, loud exit guard
- git mv tests/integration/setup-run.integration.ts -> setup-run.test.ts: bun
  only collects *.test.*/*.spec.* filenames, so the suite reported exit 0 with
  ZERO tests executed, even when the path was passed explicitly.
- Loud exit guard: beforeEach swaps process.exit for a throwing stub (restored
  in afterEach), so the wizard's bail() -> process.exit(0) can never silently
  kill the test process with a green exit code again.
- Rebuilt all 9 tests' @clack/prompts answer queues against the wizard's REAL
  prompt order (captured empirically): recommended-defaults confirm first,
  then stash dir, embedding/LLM, semantic search, registries, registry
  stashes, add-source loop, output, server-install + scheduled tasks,
  small-model provider, agent connection, save. No assertion weakened.
- Strict queue mock: drained queue throws with the full prompt trace;
  SETUP_RUN_CAPTURE=1 re-captures the sequence after future flow changes.
- Hoisted the 9 duplicated @clack/prompts mock.module blocks into one
  file-local installClackMock() (per the design doc's deferred-@clack section;
  @clack keeps mock.module by design) and collapsed the per-test seam stanzas
  into installSetupSeams(), steered via setupState.
- Fixed the stale indexer/db mock to export openIndexDatabase (the name
  setup.ts actually imports; the old openDatabase mock left it undefined so
  the vec-probe test exercised a TypeError, not the real "db locked" path).
- Seamed loadSetupStashes to a fixed list — the real loader fetches the live
  registry index, which made the stash multiselect options nondeterministic.
- Updated references to the old filename (release-check.sh, docs).
